fix: re-use existing connection info during force refresh
When the connector initiates a force refresh, it would previously
discard existing connection info and block on the refresh operation
completing. In case the cause of the connection failure is transient, we
keep the current connection info around and refresh in the background.
When the current connection info is in fact invalid, this will cause
callers to see additional failures until the refresh completes (usually
< 1s).